add logic to remove derive state in peopleToMerge and peopleNoToMerge lists, add onDeselect and onSelect functions to columns schema, remove unused getColumns function
